requirements.txt
streamlit
pandas
matplotlib
python-dotenv
requests
